Run forty-three more suites on the engine the box now runs
A scan of every SQL template in the repository found PostgreSQL-only
spellings still sitting in fifty-two files. A statement carrying one is
refused before it runs, so the case around it dies in setup and asserts
nothing — a suite in that state is green-looking silence, not coverage.
Four of those files are shipping product code, and one of them is the
monthly reconciliation against the Spanish tax agency.

The product code first.

The reconciliation picked the year, the month and the display form of an
invoice's expedition date with three calls to `to_char`, which this engine
does not have, so the whole monthly audit could not run. The date is stored
as text in year-month-day order, so the filter now compares the first seven
characters against one value of that shape, and the display form goes
through the one function this package already owns for turning a stored
date into the tax agency's own order — the exact reverse of what wrote the
column. Getting a month or a year boundary wrong here is an accounting
error, so it was measured rather than reasoned about: rows on the last day
of December, the first day of January, a single-digit month and a
deliberately malformed date were written and queried, and the control in
the other direction (a month prefix matching nothing against a table that
holds that month) was run too. A single-digit month is stored zero-padded,
and an unpadded query returns nothing rather than quietly matching.

An acknowledgement write cast a timestamp; both sides of that fallback are
already text in the same format, so the cast had nothing left to do. Two
counts were cast to text because the old driver returned a count as a large
integer object the caller then parsed. This engine returns a plain number,
measured both for three rows and for none, so the cast and the parsing
around it are gone and the type says `number` where it had been claiming
`string`.

Then thirty-nine test files across seven packages. Four kinds of change
account for nearly all of it: a cast that existed only to flatten a count;
a read of PostgreSQL's catalogue tables, replaced by this engine's own
table-shape functions following an idiom already in the tree; an array or
JSON cast on a column that is plain text now; and a `truncate` in a setup
hook, which is not a statement this engine has at all and was killing whole
files before their first assertion — one of them all twenty-nine cases.

A second blocker turned up underneath that one and is worth naming because
no text scan can see it: a hand-written insert that omits a column whose
default moved from the database into JavaScript is refused outright. It
cost six of one seat's thirteen files. Only running the suites finds it.

Sixteen cases were deleted, each with what it held written down beside the
suite it left. They fall into three groups: cases asking about roles,
grants or a search path on an engine that has none; cases asserting that a
`TRUNCATE` is refused, where the statement does not parse and the
append-only protection is proven elsewhere by attempting a real update and
a real delete; and one case about word-stemming dictionaries that no longer
exist. Two append-only cases that read a trigger catalogue were retired in
favour of the root guard that tries the offending write on every ledger
table by name, which is the stronger check.

Several assertions lost a half, and each is annotated where it stands
rather than quietly narrowed: a column's type name, a numeric precision and
scale, an enum type name, and two foreign keys pinned by a constraint name
this engine does not give them. No fiscal value moved — every hash, amount
and refusal asserted in the immutability and money-column suites is
byte-for-byte what it was.

Measured at the commit, diffed by test name against a run with every one of
these changes removed: 1,295 passing cases to 1,494 across the seven
packages — `db` 388 to 449, `fiscal-verifactu` 275 to 342, `payments` 250 to
285, `venue-service` 51 to 79, `identity` 215 to 222, `fiscal-none` 11 to 12,
`media` unchanged at 105 with one failing case deleted. **Zero passing cases
lost in any of the seven.** Root project 3,246 passing and the two
deliberate reds already recorded in the previous commit, nothing else. Lint
and format clean.

Six things were found and deliberately not fixed; all six are in the
handoff notes and the open-questions file, and the first is live and
fiscal: the drainer hands this engine a value it will not accept, the
drainer's own catch swallows it, and so every submission silently rolls
back and no record is ever marked accepted.

+++ packages/fiscal-verifactu/src/restore.pg.test.ts
@@ -35,8 +35,12 @@ describe("restoreFiscal", () => {
     expect(after.id).not.toBe(before.id);
     expect(after.numeroInstalacion).toBeGreaterThanOrEqual(installationFloor(NOW));
     expect(await withTransaction(suite.db, (tx) => esPrimerRegistro(tx, till.nodeId))).toBe(true);
+    // `count(*)::int` was casting away the BigInt the PostgreSQL driver returned. Measured on
+    // node v26.7.0 against `node:sqlite`: `select count(*) as n` comes back as a JavaScript
+    // `number` (`typeof` is `"number"`), so nothing needs casting — and the `::` the cast needed
+    // reaches this engine as `unrecognized token: ":"`, the failure this file opened with.
     const { rows: ledger } = await suite.db.execute<{ n: number }>(
-      sql`select count(*)::int as n from registros_facturacion where node_id = ${till.nodeId}`,
+      sql`select count(*) as n from registros_facturacion where node_id = ${till.nodeId}`,
     );
     expect(ledger[0]?.n).toBe(1);
     expect(outcome.series?.map((s) => s.code)).toEqual([`GA-${after.numeroInstalacion}`]); // seedTill's `GA` standard series
