+++ packages/identity/src/staff.test.ts
@@ -43,8 +43,12 @@ function run<T>(fn: (tx: Transaction) => Promise<T>): Promise<T> {
   return withTransaction(suite.db, fn);
 }
 
+// No cast on the count. The `::int` this carried was refused before the statement ran —
+// `unrecognized token: ":"`, because a colon opens a bind parameter to SQLite's parser. It was
+// there to turn the PostgreSQL driver's BigInt into a number; measured on node v26.7.0, this driver
+// hands `select count(*)` back as a JavaScript number already (`3`, `typeof "number"`).
 async function personCount(): Promise<number> {
-  const rows = await suite.db.execute<{ n: number }>(sql`select count(*)::int as n from persons`);
+  const rows = await suite.db.execute<{ n: number }>(sql`select count(*) as n from persons`);
   return rows.rows[0]!.n;
 }
